pyproject.toml
[tool.poetry]
name = "bookmarks"
version = "0.1.0"
description = ""
authors = ["Jochem Gietema"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "*"
python-dotenv = "*"
click = "*"

[tool.poetry.dev-dependencies]
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = "\\.pyi?$"
exclude = """
/(
    \\.git
  | \\.tox
  | \\.venv
  | notebooks
  | reports
)/
"""

